requirements.txt
pypng==0.0.20
bitstring==3.1.7